feat(module): changes the return to an object
The SpelunkerModule now returns an array of objects that represent
each module registered in the application. This is to allow for more
programmatic use of the module instead of direct console logging.

BREAKING CHANGES: module no longer accepts a logger parameter
and now returns an array of objects instead of void.