feat(module): supports dynamic modules
The SpelunkerModule now supports the use of dynamic modules
and can show the imports and exports of them along with
their dependencies. The only caveat is dynamic modules
that also make use of the `@Module()` decorator. Those
are still to be a work in progress.

BREAKING CHANGE: The `debug` method is now asynchronous.
This is due to the fact of needing to resolve promise
based imports.