feat: general code update for Nest v8
I've moved to using uvu as a test runner instead of jest and pnpm as a
package manager instead of yarn. I'll add in changeset as well before
publishing and set up tests for circular dependencies for both explore
and debug as those are very useful things to try and take care of for
a module like this.